+++ src/10_adjustmentsTrailer.h
@@ -11,15 +11,17 @@
 
 // IMPORTANT!! Replace the addresses below with your trailers MAC addresses!! --------------------------
 // Trailer 1
-uint8_t broadcastAddress1[] = {0x4C, 0xEB, 0xD6, 0x7C, 0x02, 0x74}; // Bruder Low Loader
+//uint8_t broadcastAddress1[] = {0x4C, 0xEB, 0xD6, 0x7C, 0x02, 0x74}; // Bruder Low Loader
+uint8_t broadcastAddress1[] = {0xAC, 0x67, 0xB2, 0x37, 0xF3, 0x7C}; // SCANIA S1  (this address is causing engine revving issues, if used in combination with the trailer 2 Scania address!)
 //uint8_t broadcastAddress1[] = {0x4C, 0xEB, 0xD6, 0x7C, 0x1E, 0x10}; // US flatbed
 //uint8_t broadcastAddress1[] = {0xAC, 0x67, 0xB2, 0x12, 0x30, 0x28}; // white board
 //uint8_t broadcastAddress1[] = {0xA0, 0x20, 0xA6, 0x10, 0x46, 0x3B}; // D1 Mini ESP8266
 
 // Trailer 2
-//#define TRAILER_2 // Uncomment this, if you want to use a 2nd trailer
-//uint8_t broadcastAddress2[] = {0x4C, 0xEB, 0xD6, 0x7C, 0x02, 0x74}; // Bruder Low Loader
-uint8_t broadcastAddress2[] = {0xA0, 0x20, 0xA6, 0x10, 0x46, 0x3B}; // D1 Mini ESP8266
+#define TRAILER_2 // Uncomment this, if you want to use a 2nd trailer
+uint8_t broadcastAddress2[] = {0x4C, 0xEB, 0xD6, 0x7C, 0x02, 0x74}; // Bruder Low Loader
+//uint8_t broadcastAddress2[] = {0x34, 0x86, 0x5D, 0x3A, 0xF1, 0x5C}; //SCANIA S2
+//uint8_t broadcastAddress2[] = {0xA0, 0x20, 0xA6, 0x10, 0x46, 0x3B}; // D1 Mini ESP8266
 
 // Trailer 3 (EXPERIMENTAL, causing engine RPM issues!)
 //#define TRAILER_3 // Uncomment this, if you want to use a 3rd trailer