+++ src/3_adjustmentsESC.h
@@ -109,7 +109,7 @@ const uint16_t escTakeoffPunch = 0;
 const uint16_t escReversePlus = 0;
 
 // Brake margin: (Experimental!)
-// This setting prevents the ESC from going completely back to zero / neutral as long as the braker trigger is pulled.
+// This setting prevents the ESC from going completely back to zero / neutral as long as the brake trigger is pulled.
 // This prevents the vehicle from rolling back as long as brake is applied. 0 = no effect, ca. 20 = strong effect.
 // How it works? Prevents the ESC from entering the "drag brake range"
 // Warning: vehicle may be unable to stop, if too high, especially when driving downhill! NEVER more than 20!
@@ -119,11 +119,11 @@ const uint16_t brakeMargin = 0; // For RZ7886 motor driver and 370 motor = 10
 const uint8_t crawlerEscRampTime = 10; // about 10 (15 for Jeep), less = more direct control = less virtual inertia
 
 // Allows to scale vehicle file dependent acceleration
-uint16_t globalAccelerationPercentage = 70; // about 100 - 200% (200 for Jeep) Experimental, may cause automatic transmission shifting issues!
+uint16_t globalAccelerationPercentage = 150; // about 100 - 200% (200 for Jeep, 150 for 1/8 Landy) Experimental, may cause automatic transmission shifting issues!
 
 /* Battery low discharge protection (only for boards with voltage divider resistors):
 *  IMPORTANT: Enter used resistor values in Ohms (Ω) and THEN adjust DIODE_DROP, until your readings match the actual battery voltage! */
-//#define BATTERY_PROTECTION // This will disable the ESC output, if the battery cutout voltage is reached
+#define BATTERY_PROTECTION // This will disable the ESC output, if the battery cutout voltage is reached
 const float CUTOFF_VOLTAGE = 3.3; // Usually 3.3 V per LiPo cell. NEVER below 3.2 V!
 const float FULLY_CHARGED_VOLTAGE = 4.2; // Usually 4.2 V per LiPo cell, NEVER above!
 const float RECOVERY_HYSTERESIS = 0.2; // around 0.2 V
@@ -131,9 +131,9 @@ const float RECOVERY_HYSTERESIS = 0.2; // around 0.2 V
  * When selecting resistors, always use two of the same magnitude: Like, for example, 10k/2k, 20k/4k or 100k/20k. NEVER exceed a ratio LOWER than (4:1 = 4)!
  * WARNING: If the ratio is too LOW, like 10k/5k (2:1 = 2), the battery voltage will most likely DAMAGE the controller permanently!
  * Example calculation: 2000 / (2000 + 10000) = 0.166 666 666 7; 7.4 V * 0.167 = 1.2358 V (of 3.3 V maximum on GPIO Pin). */
-uint32_t RESISTOR_TO_BATTTERY_PLUS = 99900; // Value in Ohms (Ω), for example 10000
-uint32_t RESISTOR_TO_GND = 19800; // Value in Ohms (Ω), for example 2000. Measuring exact resistor values before soldering, if possible is recommended!
-float DIODE_DROP = 0.0; // Fine adjust measured value and/or consider diode voltage drop (about 0.34V for SS34 diode)
+uint32_t RESISTOR_TO_BATTTERY_PLUS = 9600; // Value in Ohms (Ω), for example 10000
+uint32_t RESISTOR_TO_GND = 1000; // Value in Ohms (Ω), for example 2000. Measuring exact resistor values before soldering, if possible is recommended!
+float DIODE_DROP = 0.31; // Fine adjust measured value and/or consider diode voltage drop (about 0.34V for SS34 diode)
 /* It is recommended to add a sticker to your ESP32, which includes the 3 calibration values above */
 volatile int outOfFuelVolumePercentage = 80; // Adjust the message volume in %
 // Select the out of fuel message you want: