+++ src/7_adjustmentsServos.h
@@ -21,8 +21,9 @@
  
  // Select the vehicle configuration you have:
 //#define SERVOS_DEFAULT // <------- Select (remove //) one of the remote configurations below
-//#define SERVOS_LANDY
-#define SERVOS_C34
+//#define SERVOS_LANDY_MN_MODEL
+#define SERVOS_LANDY_DOUBLE_EAGLE
+//#define SERVOS_C34
 //#define SERVOS_URAL
 //#define SERVOS_RGT_EX86100
 //#define SERVOS_ACTROS
@@ -49,7 +50,7 @@ const uint16_t STEERING_RAMP_TIME = 0; // 0 = fastest speed, enlarge it to aroun
 
 #endif
 
-// Land Rover Defender servo configuration profile -------------------------------------------------------------------------------------------
+// MN Model 1:12 Land Rover Defender servo configuration profile -------------------------------------------------------------------
 #ifdef SERVOS_LANDY
 
 // Servo frequency
@@ -61,6 +62,28 @@ const uint8_t SERVO_FREQUENCY = 50; // usually 50Hz, some servos may run smoothe
 const uint16_t CH1L = 1840, CH1C = 1485, CH1R = 1090; // CH1 steering left 1880, center 1480, right 1080
 const uint16_t CH2L = 978, CH2C = 1833, CH2R = 1833; // CH2 transmission gear 1 978, 2 1833, 3 1833
 const uint16_t CH3L = 1300, CH3C = 1450, CH3R = 1600; // CH3 winch pull, off, release
+const uint16_t CH4L = 1300, CH4R = 1700; // CH4 trailer coupler (5th. wheel) locked, unlocked
+
+// Servo ramp time 
+const uint16_t STEERING_RAMP_TIME = 0; // 0 = fastest speed, enlarge it to around 3000 for "scale" servo movements
+
+#endif
+
+// Double Eagle 1:8 Land Rover Defender servo configuration profile -------------------------------------------------------------------
+#ifdef SERVOS_LANDY_DOUBLE_EAGLE
+
+#define MODE2_WINCH // Mode 2 is used for winch mode, if defined. The winch is controlled by the CH4 pot and connected to Servo CH3. BUS mode only!
+//#define NO_WINCH_DELAY // Use this, if you don't want a winch on / off ramp
+
+// Servo frequency
+const uint8_t SERVO_FREQUENCY = 50; // usually 50Hz, some servos may run smoother @ 100Hz
+ 
+// WARNING: never connect receiver PWM signals to the "CH" pins in BUS communication mode!
+
+// Servo limits 
+const uint16_t CH1L = 900, CH1C = 1600, CH1R = 2200; // CH1 steering left 900, center 1600, right 2200
+const uint16_t CH2L = 1900, CH2C = 1000, CH2R = 1000; // CH2 transmission gear 1 1900, 2 1000, 3 1000
+const uint16_t CH3L = 2000, CH3C = 1500, CH3R = 1000; // CH3 winch pull, off, release
 const uint16_t CH4L = 1300, CH4R = 1700; // CH4 trailer coupler (5th. wheel) locked, unlocked
 
 // Servo ramp time 
@@ -110,6 +133,7 @@ const uint16_t STEERING_RAMP_TIME = 0; // 0 = fastest speed, enlarge it to aroun
 #ifdef SERVOS_RGT_EX86100
 
 #define MODE2_WINCH // Mode 2 is used for winch mode, if defined. The winch is controlled by the CH4 pot and connected to Servo CH3. BUS mode only!
+#define NO_WINCH_DELAY // Use this, if you don't want a winch on / off ramp
 
 // Servo frequency
 const uint8_t SERVO_FREQUENCY = 50; // usually 50Hz, some servos may run smoother @ 100Hz